+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### 2023-09-12
+* Async texture fetch support (`target.read(box, data=>{...})`)
+
 ### 2023-09-07
 * Samplers support (see [TextureSamplers](https://google.github.io/swissgl/#TextureSamplers) demo)